+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   `fun :bar -2 2` was parsed as operator call for `-`.
   Sometime in the future we should change to a proper tokenizer.
 * Bugfix: Assigning to a local variable cleared it's :ref state.
+* Feature: Added optional serde/serde_json implementation and prelude
+function `ser:json` and `deser:json`.
 
 0.2.1 (2019-07-18)
 ==================